Move VenueDetailsBlock to blocks.py
